feat(kernel): establish governed edge runtime boundaries
Extract the Energy Pack and runtime manifest contracts, enforce application command boundaries and typed SHM authority, and align CLI, MCP, Swagger, installers, and documentation with the industry-neutral kernel distribution.

+++ services/automation/src/api/routing_query_handlers.rs
@@ -24,7 +24,7 @@ use crate::error::AutomationError;
     get,
     path = "/api/instances/{id}/routing",
     params(
-        ("id" = u16, Path, description = "Instance ID")
+        ("id" = u32, Path, description = "Instance ID")
     ),
     responses(
         (status = 200, description = "Instance routing categorized by type", body = serde_json::Value,
